+++ packages/papupata/src/test/typed-queries/boolean-test.ts
@@ -24,7 +24,7 @@ describe('typed-queries/boolean', function () {
   describe('single', function () {
     it('client, value true', async function () {
       const path = getUniquePath()
-      const api = API.declareGetAPI(path).query({ q1: Boolean }).response<string>()
+      const api = API.declareGetAPI(path, undefined, { disableAutoImplement: true }).query({ q1: Boolean }).response<string>()
       testServer.getApp().get(path, (req, res) => res.send('Value: ' + req.query.q1))
 
       // When
@@ -36,7 +36,7 @@ describe('typed-queries/boolean', function () {
 
     it('client, value false', async function () {
       const path = getUniquePath()
-      const api = API.declareGetAPI(path).query({ q1: Boolean }).response<string>()
+      const api = API.declareGetAPI(path, undefined, { disableAutoImplement: true }).query({ q1: Boolean }).response<string>()
       testServer.getApp().get(path, (req, res) => res.send('Value: ' + req.query.q1))
 
       // When
@@ -119,7 +119,7 @@ describe('typed-queries/boolean', function () {
   describe('array', function () {
     it('client', async function () {
       const path = getUniquePath()
-      const api = API.declareGetAPI(path)
+      const api = API.declareGetAPI(path, undefined, { disableAutoImplement: true })
         .query({ q1: [Boolean] })
         .response<string>()
       testServer.getApp().get(path, (req, res) => res.send('Value: ' + req.query.q1.join('*')))