+++ packages/papupata/src/test/typed-queries/number-test.ts
@@ -24,7 +24,9 @@ describe('typed-queries/number', function () {
   describe('single', function () {
     it('client', async function () {
       const path = getUniquePath()
-      const api = API.declareGetAPI(path).query({ q1: Number }).response<string>()
+      const api = API.declareGetAPI(path, undefined, { disableAutoImplement: true })
+        .query({ q1: Number })
+        .response<string>()
       testServer.getApp().get(path, (req, res) => res.send('Value: ' + req.query.q1))
 
       // When
@@ -131,7 +133,7 @@ describe('typed-queries/number', function () {
   describe('array', function () {
     it('client', async function () {
       const path = getUniquePath()
-      const api = API.declareGetAPI(path)
+      const api = API.declareGetAPI(path, undefined, { disableAutoImplement: true })
         .query({ q1: [Number] })
         .response<string>()
       testServer.getApp().get(path, (req, res) => res.send('Value: ' + req.query.q1.join('*')))